+++ bin/plugins/checkPlugin.js
@@ -4,10 +4,10 @@
 *
 * Usage -- see README.md
 *
-* Normal usage:                node bin/plugins/checkPlugins.js ep_whatever
-* Auto fix the things it can:  node bin/plugins/checkPlugins.js ep_whatever autofix
+* Normal usage:                node bin/plugins/checkPlugin.js ep_whatever
+* Auto fix the things it can:  node bin/plugins/checkPlugin.js ep_whatever autofix
 * Auto commit, push and publish(to npm) * highly dangerous:
-node bin/plugins/checkPlugins.js ep_whatever autofix autocommit
+node bin/plugins/checkPlugin.js ep_whatever autofix autocommit
 
 */
 
@@ -459,7 +459,7 @@ fs.readdir(pluginPath, (err, rootFiles) => {
       `cd node_modules/${pluginName}`,
       'git rm -rf node_modules --ignore-unmatch',
       'git add -A',
-      'git commit --allow-empty -m "autofixes from Etherpad checkPlugins.js"',
+      'git commit --allow-empty -m "autofixes from Etherpad checkPlugin.js"',
       'git push',
       'cd ../..',
     ].join(' && ');