+++ CHANGELOG.md
@@ -40,7 +40,7 @@
   content in `.etherpad` exports
 * New `expressCloseServer` hook to close Express when required
 * The `padUpdate` hook context now includes `revs` and `changeset`
-* `checkPlugins.js` has various improvements to help plugin developers
+* `checkPlugin.js` has various improvements to help plugin developers
 * The HTTP request object (and therefore the express-session state) is now
   accessible from within most `eejsBlock_*` hooks
 * Users without a `password` or `hash` property in `settings.json` are no longer